fix: get TIMESERIES_TABLE_NAME from env, rename types.py to fix circular import
