+++ space2stats_api/src/space2stats/model_types.py
