+++ space2stats_api/src/space2stats/settings.py
@@ -8,6 +8,7 @@ class Settings(BaseSettings):
     PGUSER: str
     PGPASSWORD: str
     PGTABLENAME: str
+    TIMESERIES_TABLE_NAME: str
 
     # see https://www.psycopg.org/psycopg3/docs/api/pool.html#the-connectionpool-class for options
     DB_MIN_CONN_SIZE: int = 1
